% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_workflow.R
\name{example_workflow}
\alias{example_workflow}
\title{Complete Workflow Example for embeddcv Package}
\usage{
example_workflow()
}
\description{
This example demonstrates a complete workflow using the embeddcv package,
from loading data to creating Sankey visualizations.
}
\examples{
\dontrun{
# Complete workflow example - to be implemented
# example_workflow()
}

}
