% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_hf.R
\name{get_response_hf}
\alias{get_response_hf}
\title{Generate Responses using HuggingFace}
\usage{
get_response_hf(
  prompt = NULL,
  model = "google/gemma-2-2b-it",
  api_key = NULL,
  max_tokens = 128,
  temperature = 1
)
}
\arguments{
\item{prompt}{Character. The prompt to be sent to the ChatBot.}

\item{model}{Character. The embedding model to use. Defaults to 'google/gemma-2-2b-it'.}

\item{api_key}{Character. Your API key. If not provided, will try to get from environment variables.}

\item{max_tokens}{Integer. The maximum number of tokens for the generated response. Defaults to 128.}

\item{temperature}{Numeric between 0 and 1. The level of randomness of the response. Defaults to 1.}
}
\value{
A string with the ChatBot response.
#'
}
\description{
Generate a response given a prompt using HuggingFace API.
}
\details{
The function automatically determines the appropriate API key from the "HF_API_KEY" environment variable.
}
\examples{
\dontrun{
 # Get a response from Gemma (requires API key)
 Response <- get_response_hf("What is the most famous personality theory?")
 # Print the response nicely
 cat(Response)
}

}
