% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embeddings_hf_local.R
\name{check_hf_models_local}
\alias{check_hf_models_local}
\title{Check Available HuggingFace Models Locally}
\usage{
check_hf_models_local()
}
\value{
A list with information about available models and Python environment.
}
\description{
This function checks which HuggingFace models are available locally
and provides information about installed models.
}
\examples{
\dontrun{
 # Check what's available
 model_info <- check_hf_models_local()
 print(model_info)
}

}
